Move tinymce to webpacker and upgrade to v6 [SCI-7081]
