+++ app/assets/stylesheets/tiny_mce.scss
@@ -56,6 +56,11 @@
   }
 }
 
+.tinymce-save-controls {
+  display: flex;
+  margin-left: auto !important;
+}
+
 .tinymce-status-badge {
   color: $color-silver-chalice;
   margin-top: -20px;
@@ -69,18 +74,17 @@
   background: $color-white !important;
 }
 
-
-
 .restore-draft-notification {
   align-items: center;
   background: $state-info-bg !important;
   display: flex;
+  flex-basis: 100%;
   height: 30px !important;
-  padding: 0 10px !important;
+  padding: 10px !important;
 
   .notification-text {
     flex-grow: 1;
-    max-width: 75%;
+    max-width: 85%;
     overflow: hidden;
     text-overflow: ellipsis;
   }
@@ -191,6 +195,11 @@
       color: $color-volcano;
     }
   }
+}
+
+// fix for TinyMCE 6 vs Boostrap 3 .show conflict
+.tox.tox-tinymce.show {
+  display: flex !important;
 }
 
 // scss-lint:enable ImportantRule
