+++ app/javascript/packs/tiny_mce.scss
@@ -1,0 +1,2 @@
+@import 'tinymce/skins/ui/tinymce-5/skin.min.css';
+@import 'tinymce/skins/ui/tinymce-5/content.min.css';
