+++ app/views/shared/_tiny_mce_extra_buttons.html.erb
@@ -1,10 +1,10 @@
-<div class="hidden tinymce-cancel-button mce-widget mce-btn mce-menubtn mce-flow-layout-item mce-btn-has-text pull-right" tabindex="-1">
+<div class="hidden tinymce-cancel-button tox-mbtn" tabindex="-1">
   <button type="button" tabindex="-1">
     <span class="fas fa-times"></span>
     <span class="mce-txt"><%= t('general.cancel') %></span>
   </button>
 </div>
-<div class="hidden tinymce-save-button mce-widget mce-btn mce-menubtn mce-flow-layout-item mce-btn-has-text mce-last pull-right" tabindex="-1">
+<div class="hidden tinymce-save-button tox-mbtn" tabindex="-1">
   <button type="button" tabindex="-1" >
     <span class="fas fa-check"></span>
     <span class="mce-txt"><%= t('general.save') %></span>
