+++ package.json
@@ -40,19 +40,20 @@
     "flow-typed": "^2.5.1",
     "prettier": "^1.13.7",
     "prettier-eslint": "^8.8.2",
-    "webpack-dev-server": "^3.7.2"
+    "webpack-dev-server": "^4.10.0"
   },
   "dependencies": {
     "@babel/core": "^7.0.0",
     "@babel/plugin-proposal-class-properties": "^7.0.0",
     "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
     "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.18.10",
     "@babel/polyfill": "^7.0.0",
     "@babel/preset-env": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
     "@fortawesome/fontawesome-free": "^5.2.0",
     "@joeattardi/emoji-button": "^4.6.2",
-    "@rails/webpacker": "^4.0.7",
+    "@rails/webpacker": "^5.4.3",
     "autoprefixer": "^7.2.6",
     "axios": "0.21.2",
     "babel-loader": "^8.0.0",
@@ -109,6 +110,7 @@
     "shortid": "^2.2.12",
     "style-loader": "^0.18.2",
     "styled-components": "^2.4.1",
+    "tinymce": "^6.1.2",
     "tui-code-snippet": "^1.5.0",
     "tui-color-picker": "^2.2.0",
     "tui-image-editor": "github:biosistemika/tui.image-editor#3_15_2_updated",
